+++ archive/archive.R
@@ -162,7 +162,31 @@ arch_fixed_episodes <- function(x, strata = NULL, case_length, episodes_max = In
   return(x)
 }
 
-mths <- c("across", "inbetween", "chain", "reverse", "aligns_start", "aligns_end", "exact", "overlap", "none")
+# Up to v.0.4.0
+# mths <- c("across", "inbetween", "chain", "reverse", "aligns_start", "aligns_end", "exact", "overlap", "none")
+# mths <- sort(mths)
+# mths_lst <- lapply(seq_len(length(mths)), function(i){
+#   combn(length(mths), i, simplify = FALSE)
+# })
+# mths_lst <- unlist(mths_lst, recursive = FALSE, use.names = FALSE)
+# mths_lst <- lapply(mths_lst, function(x){
+#   paste0(mths[x], collapse ="|")
+# })
+# mths_lst <- unlist(mths_lst, recursive = TRUE, use.names = FALSE)
+# mths_lst_cd <- lapply(mths, function(x){
+#   if(!x %in% c("overlap", "none")){
+#     which(grepl(x, mths_lst) & !grepl("overlap|none", mths_lst))
+#   }else if (x == "overlap"){
+#     which(grepl(x, mths_lst) & !grepl("none", mths_lst))
+#   }else if (x == "none"){
+#     which(grepl(x, mths_lst))
+#   }
+# })
+# names(mths_lst_cd) <- mths
+# overlap_methods <- list(options = mths_lst, methods = mths_lst_cd)
+# save(list = "overlap_methods", file = "data/overlap_methods.RData")
+
+mths <- c("across", "inbetween", "chain", "reverse", "aligns_start", "aligns_end", "exact", "none", "overlap")
 mths <- sort(mths)
 mths_lst <- lapply(seq_len(length(mths)), function(i){
   combn(length(mths), i, simplify = FALSE)
@@ -171,16 +195,29 @@ mths_lst <- unlist(mths_lst, recursive = FALSE, use.names = FALSE)
 mths_lst <- lapply(mths_lst, function(x){
   paste0(mths[x], collapse ="|")
 })
-mths_lst <- unlist(mths_lst, recursive = TRUE, use.names = FALSE)
-mths_lst_cd <- lapply(mths, function(x){
-  if(!x %in% c("overlap", "none")){
-    which(grepl(x, mths_lst) & !grepl("overlap|none", mths_lst))
-  }else if (x == "overlap"){
-    which(grepl(x, mths_lst) & !grepl("none", mths_lst))
-  }else if (x == "none"){
-    which(grepl(x, mths_lst))
-  }
+mths_lst_1 <- unlist(mths_lst, recursive = TRUE, use.names = FALSE)
+
+mths <- c("x_across_y", "y_across_x", "x_inbetween_y",
+          "y_inbetween_x", "x_chain_y", "y_chain_x",
+          "reverse", "aligns_start", "aligns_end",
+          "exact", "none", "overlap")
+mths <- sort(mths)
+mths_lst <- lapply(seq_len(length(mths)), function(i){
+  combn(length(mths), i, simplify = FALSE)
 })
-names(mths_lst_cd) <- mths
+mths_lst <- unlist(mths_lst, recursive = FALSE, use.names = FALSE)
+mths_lst <- lapply(mths_lst, function(x){
+  paste0(mths[x], collapse = "|")
+})
+mths_lst_2 <- unlist(mths_lst, recursive = TRUE, use.names = FALSE)
+mths_lst <- c(mths_lst_1, mths_lst_2)
+mths_lst[grepl("overlap", mths_lst)] <- "overlap"
+mths_lst[grepl("none", mths_lst)] <- "none"
+mths_lst <- mths_lst[!duplicated(mths_lst)]
+
+mths_lst_cd <- lapply(c(mths, "across", "chain", "inbetween"), function(x){
+  which(grepl(x, mths_lst))
+})
+names(mths_lst_cd) <- c(mths, "across", "chain", "inbetween")
 overlap_methods <- list(options = mths_lst, methods = mths_lst_cd)
 save(list = "overlap_methods", file = "data/overlap_methods.RData")